Block cache bug fix
1) After a block is evicted, the block becomes free block. When the free
block is re-cached, it should be inserted into clean LRU list or dirty RB-tree.
There was a bug that the flag of freed block is not cleared so that the block
is neither inserted into list nor tree. This causes the corruption of list and
segmentation fault. The bug is fixed by clearing the flag when a block is freed.

2) Extra white space is removed in some files.