MB-10762 Return deleted items from fdb_get_meta and fdb_iterator
This change adapts fdb_get_meta and fdb_get_meta_byseq APIs, so
that they return a deleted item to the caller, but mark it as
"deleted". If a deleted item is purged permanently by the
compactor, then key_not_found will be returned.

In addition, fdb_iterator API is adapted to return a deleted item
optionally.

Change-Id: I5603fa653ef49d855f944754e8b03e78d2be3e96