MB-10695 Add API to read a doc body given its doc offset.
This change adds a new API to read a doc body through its
doc offset. To support this, it is required to adapt
fdb_get_metaonly, fdb_get_metaonly_byseq, and
fdb_iterator_next_offset, so that they return the offset
to a doc (header + key + metdata + body) because a doc body
can be compressed using snappy.

Change-Id: I06eb5cf409add9deb987ae4d0c935f8748543de0