MB-11213 Allow multiple nodes in a block for small B+trees
- When B+tree is created at first so that it consists of a single
unique node, a small sub-block (whose size is 128 bytes by default)
is allocated for the B+tree node, instead of a whole block.

- As B+tree grows, the size of its (unique) node also gets doubled
(128 -> 256 -> ..). The B+tree does not use small-sized block anymore
when its height becomes equal to or greater than 2.

- Multiple same-sized sub-blocks are packed in a single block so that
we can save the space occupied by ForestDB up to 1/32 when a block
size is 4096 bytes (128/4096), if there are lots of small B+trees as
leaf nodes of HB+trie.

Change-Id: I4ec8d181dc90be4d994aff60f7b32c127fb6b3b3