Update the seq number field for read-only handles.
If a given handle is purely used for read-only operations,
its seq number field should be synchronized with the current
seq number of the corresponding file.

Note that it is not strictly necessary for a read-only handle
to grab the lock to get the current seq number of a given
database file.

Change-Id: I14fa1fe7098c9dca478ab221e1fefdf15dde4e60