Flush WAL before compaction is done
- The current compaction uses WAL to move documents from old file to
new file, so that memory consumption increases temporarily during the
compaction. This can be a large overhead in resource-limited
environment such as mobile devices. For this reason, if
'wal_flush_before_commit' is set, WAL can be flushed also while
compaction is being performed. It makes the memory consumption of
ForestDB constant, but the compaction performance degrades by 1.2x-1.5x.