Fix a bug that mis-copies existing common prefix
- When a new B+tree is created between a parent B+tree and its child
B+tree, the common prefix stored in the child B+tree has to be
re-adjusted (shrunk), and a chunk from the old common prefix is
inserted into the new B+tree as a key.

- There was a bug that common prefix is modified first, and the key
to be inserted is made from the 'modified' common prefix, so that the
entire key becomes different from the original key.