MB-12242: support for in-memory snapshots
To support creation of in-memory snapshots without the
need to do an fdb_commit.
If FDB_SNAPSHOT_INMEM is passed as the snapshot marker
sequence number, then copy existing WAL items into AVL tree
avoiding reverse scanning of db file and on-disk WAL restoration
The snapshot created will be the latest in-memory copy of the
incoming db handle even if it means overwriting a committed item
with a copy of a dirty item from an uncommitted transaction from
which the snapshot was taken.
Change-Id: I05099e2d3931d97e64d7303ae312f7ff782886e5