MB-12977: handle lost docs by fast compactions
If past file of the current file being compacted is still open, then
documents can be lost if handles on the past file are not redirected
to the newly compacted file.
So, in the final step of fdb_compact, before the old file is removed,
filemgr hash table is scanned to find all files whose new_file link
points to the old_file. If such files are found, their headers are
redirected to point to the latest new_file.
Change-Id: I6a0a848cb1f43aaa0ba016efa7e27348a05e1747