Re-use user allocated memory in fdb_iterator_get/get_metaonly
Fewer memory allocations. Cannot pass uninitialized values to fdb_iterator_get
and fdb_iterator_getmetaonly.
If user memory is to be used, it is upto caller to ensure that the memory fits
the largest key/meta/body in the kvstore.
If caller does not know the size requirement for key/meta/body, then setting
that field to NULL in the fdb_doc will let the api allocate memory.
It will be upto caller to then free the memory appropriately,
(either using fdb_doc_free or other means)
Change-Id: Ic2bea1a1453bbd24f7c1d7ca92fe4264b7a14adc