Fix potential problem caused by chunk size
- In multi KVS mode, the first chunk for each key indicates KVS ID
that the key belongs. However, the current code simply converts from
the KVS ID (8-byte integer) to the chunk binary. This may cause
potential problem when the chunk size is not 8 bytes.

- To address this issue, we need to support generic endian-safe
conversion between KVS ID (8-byte integer) and variable length chunk.
Whatever the chunk size is, KVS ID is converted and aligned to the
chunk size. Since the minimum chunk size is 4 bytes, we can support
up to 2^32 KVS ID.

- Unlike ID-index, chunk size of the Seq-tree is fixed to 8 bytes
since the sequence number is always 8-byte integer.

Change-Id: I9ebb43287a59422a63c6237d6921ce2dd6facdab