Optimized btree_get_bnode_array() which unnecessarily calls malloc()
The malloc call in this function showed up as a hot spot in my high-
level CBL benchmark. The output is a malloced array of pointers, even
though only one pointer is ever needed.

I changed the function to instead return the single pointer directly
with no allocations. This sped up my entire benchmark by 8%!

Since the comment implies that in the future a btree might have multiple
nodes per block, I kept the old code around wrapped in an #ifdef. (If
you think this is unnecessary, I could take it out.)

Change-Id: I9522b36b3a7b3728a324093d80f389404098d7be