Use file scanning style compaction
- The current compaction scans docs using HB+trie iterator, so docs
are moved into new the file in a key order. However, since docs are
randomly scattered on the DB file, the overhead from random reads
becomes a bottleneck of the compaction throughput.

- Instead, we can largely improve the overall throughput if we
sequentially scan docs in a byte offset order, from the beginning of
the file.

- Use more large batch size (up to 32 MB, or 65536 docs) to maximize
write bandwidth.

- Verified that the performance of compaction for workloads smaller
than the RAM size is also improved.

Change-Id: Ief90ed3ce1dbc28e19f44dd7338976fd9b40225c