MB-13859: ensure kv_header is created before seqnum read
If compactor opens a file, compacts it and closes it, then the
ref count on the new file can become 0, and all entries of the file
are removed from the filemgr hash table.
Later if any operation is attempted using a handle on the old file,
the reopen hits a segmentation fault due to newly initialized
filemgr struct values for the compacted file while reading seqnum.

Change-Id: Ia03e25fda9a0b765698e7bf7237ad7fd7e162922