MB-14443 Put the limit on the buffer cache size by considering physical RAM
ForestDB client may try to configure a buffer cache whose size is greater
than the physical RAM available. We should avoid this situation and return
the appropriate error to the client.

By default, we don't allow the buffer cache to be greater than 80% of
the physical RAM on a machine.

Change-Id: I746b0941b2eda27b6ec5c7b8e3234a23370fb240