MB-14646: reset last_wal_flush_hdr_bid After delta move
Since the last stage of compaction calls fdb_sync_db_header
to retrieve the latest header in old_file, it also ends up
setting the last_wal_flush_hdr_bid in the handle to a value
from the old_file. We need to reset it after the move to ensure
that it does not carry the stale value. 

Change-Id: Ic73b6e27db9b6e5c34409a67b95116531c450689