MB-14060 Interleave compactor and normal writer through the old file's lock.
From the write-heavy tests, we observed that the compactor occasionally can't
catch up with the normal writer during the second phase of the compaction.

As a short-term solution, this change forces the compactor and writer to
interleave through the old file's lock, so that the compactor can catch up
with the writer in write-heavy use cases. We plan to address this issue
without affecting the writer performance in the future.

Change-Id: I69b25a7a7f5833089c77960e6c3e68355a66dc3f