MB-14060 Start the second phase of compaction with 20% of lock probability.
If the current probability is zero after the first phase of compaction,
then start the second phase of compaction with 20% of probability to allow
compaciton to catch up with the writer in case their throughput remains
the same approximately during the entire compaction period. Otherwise,
the compaction might not be able to catch up and run forever.

Change-Id: I043b40cc4fbfc641257a3767e7e624704e5df660