MB-14060 Make writer lock probability configurable.
This change introduces the maximum probability (range: 20% ~ 100%) for
the compactor to grab the writer's lock during each batch write in case
the writer's throughput is faster than the compactor, to make sure that
the compactor can keep pace with the writer and eventually complete the
compaction.

Note that we plan to reduce the compaction overhead significantly soon
and deprecate this parameter when it is not needed anymore.

Change-Id: I526dc2b0b7e8caa0d4c7329c7882c7502e9e9953