MB-15348 Use file access timestamp to select a victime file for cache eviction.
From the performance tests, we observed the contention on the lock that is used
to maintain the global file LRU list. This change reduces the lock conention
by using a file access timestamp to select a victim file for cache eviction.

In addition, this change uses a reader-writer spinlock to synchronize accesses
to the file list array in the buffer cache. Currently, the reader-writer spinlock
is only supported in non-Windows environments.

Our reader-writer spinlock implementation is based on the approach proposed in

https://jfdube.wordpress.com/2014/01/12/optimizing-the-recursive-read-write-spinlock/

Change-Id: I61b6916735d715b19572d484253e180f57d9b724