Move forestDB to use high-performance CRC32C (from platform if available).
Note: if libplatform is not available, e.g. not a couchbase build
then crc32c is not enabled, the in-built utils/crc32.cc is used.

This changes the underlying hashing function used for integrity
checking of ForestDB data.

When creating a new file, CRC32C is selected as the CRC and all data
is thus stored with a CRC32C derived checksum. All CRCs stored in
a ForestDB file are written using the same CRC function.

When reading a file the CRC32 mode is detected as part of reading
the header block, we check the header block CRC against each CRC mode
and use the one which produces a match (or report checksum fail).
This allows files written before this commit to be read by builds containing
this commit. The existing file is then maintained with the detected CRC,
i.e. new documents added to a CRC32 file will be written with CRC32, not
CRC32C.

A new flag is also added to the ForestDB open API that allows code to explicitly
request the 'legacy' CRC mode. This is primarily so we can programtically
perform upgrade tests. The flag is only valid for new files or legacy CRC files.
Opening a CRC32C file with the new flag results in FDB_RESULT_INVALID_ARGS.

Compaction now performs CRC upgrades.

Compacting a file always writes the new file with CRC32C. Thus
compaction will effectively upgrade files that are using CRC32.

Compaction meta-files are also written using crc32c, but read using
all known modes removing the need for changing the stored version number.

This patch removes the adler32 code which was unused and would complicate
upgrades further if ever enabled.

This patch removes 'crc_test.cc' which never made any assertions.
 - Note that crc32c is covered by IETF test vectors in the platform test suite.

This patch adds 'crc_migration.cc' which is a 'google-test' program that covers
the functionality in this commit.

Change-Id: I81a2ef481cb81feaccbbf9312be55328b76603c5