[BP] MB-16154: Add new API to trigger a callback upon a fatal error
Add a new API to forestdb: fdb_set_fatal_error_callback(). This allows
users of forestdb to optionally specify a callback to be invoked in
the event of a fatal error inside forestdb; before forestdb raises a
SIGABRT.

This replaces the previous manditory dumping of thread backtraces,
which is not suitable for all users of forestdb (it was added for GSI,
but kv-engine doesn't want or need it as it uses Google Breakpad to
catch all fatal errors.

Change-Id: I5408b597e4d1073ab78153ccb4f3d2091bb8160e