MB-16151: Drop docs that are only in WAL if inserted & deleted
In the last phase of compaction, if a bunch of new documents were
inserted and also immediately deleted, then the wal items for these
would all be WAL_ACT_REMOVE with 0 for both old_offset as well as
current offset. In this case _wal_flush_cmp would return 0 causing
the avl_insert to disregard the insert as a duplicate item.
This causes memory leaks and an enormous WAL after compaction slowing
insert rates and in-memory snapshot creations.

Change-Id: Ida10382e311bfa99f477cfe92c84a92b4793676a