MB-16249: fix crash during _fdb_open race on kv_header set
When auto compaction closes file, the ref count becomes zero
and the file gets freed. Later when multiple concurrent threads
reopen the new file, they race to create the kv_header in the file.
As this section was not protected, the kv_header was not set
atomically causing some threads to attempt reading empty kv_headers,
resulting in errors and crashes in say handle->op_stats.
e2etest concurrent scan test caught this.

Change-Id: I4c9acdeeccc2f1bd9faeee39e77428031386dcc7