MB-16608 Support cross-platform encryption options.
ForestDB uses the CommonCrypto library to perform file level encryption,
but that library is specific to Apple products (iOS and OS X).
To address it, this change provides an option to use other cross-platform
libraries (OpenSSL and LibTomCrypt).

_ENCRYPTION macro can be optionally passed to specify which crypto library
is used to support database encryption and set to 'commoncrypto' on iOS and
OS X, 'openssl' for OpenSSL, or 'libtomcrypt' for LibTomCrypt.

For example,

% cmake -D_ENCRYPTION=commoncrypto path_to_build_directory

Change-Id: I740eddfc4cf0ea59ef630879d872b978ad362574