MB-16219 Remove seq numbers if corresponding stale documents are reclaimed
- Although doc A is updated to doc A' so that doc A is marked as
stale, seq numbers for both A and A' still coexists in seq-tree.

- Since all updates for seq-tree are monotonic increasing insertions
only, the number of index blocks occupied by seq-tree increases over
time even though the stale blocks generated by seq-tree updates are
reclaimed periodically.

- However, removal of seq number for doc A should be postponed since
other snapshot or reader may access to doc A through the old seq number.
Hence, when stale blocks up to a commit header are reclaimed for
reusing, we then remove the postponed seq numbers, update seq-tree,
and also reclaim the stale blocks as a result of the update of
seq-tree.

Change-Id: Icd760b4adb8822d4d38fca273c028b74253d74e4