Set the writer throttling delay in a less aggressive way.
Setting the writer's throttling delay to 4X of the compactor's delay per doc
is too much aggressive and consequently increases the writer's latency and
reduce its throughput.

From the tests, lowering the ratio improves the writer's latency and throughput
without compromising the compaction time (within 5%).
We plan to make the ratio auto-tunable as a separate improvement task.

Change-Id: I9d602b207dba39c77022761e306d30f3db01c9ca