MB-17004: Prune an old doc's seqnum from sequence index on WAL flush
If a key has been updated or deleted, its entry is removed from the
main index by key but the entry in the sequence tree/trie is not removed.
As a result, the sequence tree continues to grow indefinitely until
compaction is executed, which in turn bloats up the index node space used.
Since we do not support having duplicate values for the same key in the
same snapshot, we can prune the sequence tree to prevent file size explosion.

Change-Id: I4aa2e44261873523e41a347171fd80236d839fc0