MB-17757 Return file system's error code to the client upon operation failure
A file system's error can happen while performing various file operations and
is captured in errno in linux and _get_errno() or GetLastError on Windows.
These errors are logged through the log callback from the client, but not
returned to the client as part of a response.

To address the above issue, this commit adds a new set of error codes that
correspond to potential errno codes that can happen while performing file
operations.

Change-Id: I84fe2edd281689570798e30cee5afe054a972a90