MB-18747: Don't lock on wal snapshot close
GSI tests like stale=ok are extremely sensitive to higher
latencies in the query path. If during wal_flush the snapshots
are closed in a highly concurrent query thread, the query
thread stalls, causing go to spawn more threads and eventually
running out of threads. Simple fix is to remove the unnecessary
locking in the close path.

Change-Id: Ica601d3420eb3be099f26f0ff141967518975859
Reviewed-on: http://review.couchbase.org/61747
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>