MB-19322 Tolerate checksum error in the delta phase of compact_upto
- When we scan a block that is an intermediate block of the series of
doc blocks (that is caused by circular scanning; when former part is
written at the end of file, and the latter part is written at the
beginning of the file), reading the first doc (offset 0) of the block
will cause checksum error.

- Since the doc block is already read by the previous scanning, we
can just ignore the checksum error (please refer to the comments
in _fdb_compact_move_delta()).

Change-Id: Ie73064adfd78ee6837afd1d2c63dae9ba7310e90
Reviewed-on: http://review.couchbase.org/63397
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>