MB-16493: Introduce a new error message: FDB_RESULT_CANCELLED
This will be the return status of the fdb_changes_since API
when the caller explicitly requests the API to stop execution
through the callback. This new status message is required to
let the caller know that the API was interrupted before
completion, so the caller could take any necessary steps in
light of such an event.

Change-Id: I208d765daeacabc31e6ad9aee272d9ab8c867444
Reviewed-on: http://review.couchbase.org/64181
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>