MB-19390: Calculate time diff with PerfFreq on windows
On windows high frequency time is computed using QueryPerformanceCounter
While its output in itself seems to be correct, the value actually needs
to be computed by also considering the counter frequency which is
number of counter units in a second.
So the formula is something like
QueryPerformanceCounter() / QueryPerformanceFrequency()
This change adjusts the ts_diff() by accounting for the above.

Change-Id: I188d75d446b5beb0567e711a520e1eecb805ddde
Reviewed-on: http://review.couchbase.org/64543
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>