Set static compaction window_size in 1st phase of compaction
In _fdb_compact_move_docs, window_size was resized based on the number
of documents acquired by re-syncing the db header outside
the file lock. If commits had happened after lock was released then
the doc_count obtained in this function is not valid since it reflects
docs to be moved only later in the delta phase.
So for correctness, simply let the static window_size setting take effect.
Side effect of this change is that multiple wal flushes may be done on
the new_file only when 1st phase has more than 8M items
(Please see MB-20040 for more details)

Change-Id: I3b78ddbbddf3d8749b9f878806d00ef169b96ba6
Reviewed-on: http://review.couchbase.org/66500
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>