MB-18453: Fix incorrect waketime for executor threads
When a thread is done with a task from a particular queue,
it calls doneWork() which was setting its current Queue type
to NONE. This was causing the subsequent reschedule to not
pick the earliest waketime and end up sleeping longer than
necessary. In rare occasions, this can even lead to
starvation. Simplify the reschedule logic as part of
this fix.

Change-Id: If909c7ab30d898097932ae5db276d2f13534155e
Reviewed-on: http://review.couchbase.org/67395
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>