MB-20801: Let WAL directly use caller's key+doc memory
If FDB_DOC_MEMORY_SHARED is set in the fdb_doc, then WAL
will skip malloc and memcpy and stash the pointer to the
doc directly instead of its offset.
insert & find apis adapted to store and return caller's
pointer instead of offset if applicable.
This sets the stage for mmaped commit log entries to be
inserted re-used by the WAL directly.
+unit test to load & fetch large number of items in 2 modes

Change-Id: Iedc517491c5dd194e1aa0add576ebfd6f20c1369
Reviewed-on: http://review.couchbase.org/67464
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>