MB-20091: Log conflicting thread's func on HANDLE_BUSY
Before an error FDB_RESULT_HANDLE_BUSY is returned from an API call,
a message about the error will be logged that includes both the name
of the function being called, and (importantly!) the name of the
function currently using the handle on another thread.
This can help in tracking down the source of these errors in the
client code.
(Adapted from original proposal by Jens Alfke)

Change-Id: I6a512833d76abe377c2a6daa4226031f3c59ef62
Reviewed-on: http://review.couchbase.org/67589
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>