[BP] MB-20957 Don't select the oldest victim if its last timestamp is not too old
When an ejection is required in the buffer cache, the old implementation
simply chooses the file that has the oldest access timestamp among the files
that are randomly selected. However, this policy can cause quite aggressive
ejections in the file having the oldest timestamp if the oldest and newest
access timestamps are not significantly huge.

To resolve the above issue, this commit adds the threshold in comparing
the oldest and newest timestamps and only selects the file having the oldest
timestamp if the timestamp difference is above the threshold. Otherwise,
the file having the largest number of cached items will be selected as a
ejection victim.

(Reviewed-on: http://review.couchbase.org/67760)

Change-Id: I5f42b34838d2aa4410a70eeb0a10b603d323d96b
Reviewed-on: http://review.couchbase.org/67836
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>