MB-21953: Re-open compacted file in rollback to 0 retry
The following racy sequence causes a hang in rollback..
1. Default KV Store opened, new handle created for default kvs.
2. Compaction runs and changes file to REMOVED_PENDING
3. rollback attempted with Default KV store handle still
pointing to old file.
4. rollback to zero attempts to remove kvs but
finds current file of root handle in FILE_REMOVE_PENDING
state and retries operation.
During retry, new file is not re-opened but rather same file is
examined repeatedly causing a hang.
Simple fix is to check for file reopen on retry even for
rollback to zero
+Test case to reproduce the hang.

Change-Id: Id7e89873c7e283711ce1a1d411ee76e804e0dd06
Reviewed-on: http://review.couchbase.org/70976
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>