[BP] MB-22046: forestdb_dump --key must work on deleted keys too
Don't dump full header if a specific key or specific kv store
name is given. This is to aid testing code which is just
examining if a given key made it to the storage end.
Search by meta if search by key fails so that deleted keys
can be point searched too.

Change-Id: I460f663d5d7eab3ac89f814c12fb1be3c78c11ed
Reviewed-on: http://review.couchbase.org/79133
Reviewed-by: Jung-Sang Ahn <jungsang.ahn@gmail.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>