Guard against LIBRT=NOTFOUND on APPLE
As a result of upgrading the version of google benchmark used we need to
guard against LIBRT being set in the CMakeCache on MacOS which does not
have LIBRT. This is because google benchmark tries to find it without
checking it's not making for a non MacOS system. This means that when
it's not found it sets the variable as NOTFOUND rather than it not being
set at all. This interfered with the forestdb CMakeLists.txt.

Change-Id: I01b8ae6b041062f27eaa562722e016db1005910f
Reviewed-on: http://review.couchbase.org/73451
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Jung-Sang Ahn <jungsang.ahn@gmail.com>