MB-26165:[utils/debug.cc] Use ucontext_t instead of ucontext.
In recent glibc versions the struct ucontext was renamed to
ucontext_t. The details can be found at [1].

The change is compatible with older glibc versions because the typedef
for the struct has been called ucontext_t for a long while. So it will
be picked up instead of the struct tag.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=21457

Change-Id: I6096e1576a973e8bcb33cad3a7cc94c2009d390d
Reviewed-on: http://review.couchbase.org/83579
Reviewed-by: Sundararaman Sridharan <sundar@couchbase.com>
Tested-by: Sundararaman Sridharan <sundar@couchbase.com>