MB-26434: After a failed compaction, the original file cannot be re-compacted
This fixes a regression caused by a change to update the next file after
compaction recovery.
https://github.com/couchbase/forestdb/blob/spock/src/forestdb.cc#L2194

The fix is to not have the old filename and new filename in the handle.
Recovery of the file has made sure the correct file is now used based
on whether the compaction succeeded or failed.

Added more logging to follow the compaction recovery activity.
Added test cases to verify the re-compaction works after recovery.

Change-Id: I1ec74abb8c23000c64366a0e0e0ff1cf03e823c9
Reviewed-on: http://review.couchbase.org/85266
Reviewed-by: Sundar Sridharan <sundar@couchbase.com>
Tested-by: Srinath Duvuru <sduvuru@gmail.com>