MB-26978: Unit test failed with "fdb_anomaly_test........*** Exception
Compaction opens the new file, switches to it, closes and unlinks
upon success. However when compaction fails after the switch, the old file is
not closed. An attempt to shutdown forestdb, destroys the compactor
mutex and then fails with file busy error. The next attempt to open the
old file does not initialize the compactor mutex as forest db was not
shutdown. When an attempt to signal a condition for that mutex is made,
the signal call hangs as the associated mutex is destroyed.

The fix is to close the old file upon a failure.

The tests for compaction failure in anomaly tests are also updated to
look for the status of all the calls and validate them.

Change-Id: Ifed47c06442c18c295d493a54d266c37cb02cf71
Reviewed-on: http://review.couchbase.org/86676
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sundar Sridharan <sundar@couchbase.com>
Tested-by: Srinath Duvuru <sduvuru@gmail.com>