MB-27005: Attempting to write to a buffer still in use
sb_bmp_is_writable() was reporting a block is not writable. This is because the
test bid >= last_commit fails. This is due to the last_writable_bmp_revnum
getting set in all cases, even when last_commit was set to the end of file.
This causes the a bitmap version mismatch in the check for a writable block.
This fix is a work around to avoid the failure. By setting the last_writable_bmp_revnum
to current bitmap version only when last_commit is not set to end of file.

Change-Id: I4ac68b381599f9ebf87badb3fbb6e350203b8a8e
Reviewed-on: http://review.couchbase.org/88454
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sundar Sridharan <sundar@couchbase.com>
Tested-by: Sundar Sridharan <sundar@couchbase.com>