MB-28014: Index node crashing with FDB error
When ForestDB encounters a corrupted buffer, currently it just asserts and crashes.
When indexer restarts, it again encounters the same corrupted buffer and crashes again.
The indexer then has no way of determining the specific index file (forestDB file)
which is corrupt and disable the specific index.

This fix will handle corruption that is detected in the following way.
If buffer (on disk or on the buffer pool) is found to be corrupt,
the current dbheader will be marked as invalid and an error, FDB_RECOVERABLE_ERR,
will be returned. When the file is reopened, forestdb will move to the first valid
dbheader skipping the DBheaders that were marked as invalid. If the forestdb file open
encounters "num_keeping_headers", a config parameter, invalid dbheaders it will return
FDB_NONRECOVERABLE_ERR, a nonrecoverable error. The indexer will now be able to
determine the index is not recoverable and invalidate it so that it can be rebuilt.

The fix also has two functional tests which corrupt index and data pages and recovers
up to num_keeping_headers configuration value.

Change-Id: I181abface1ffe2e926597dccb883bb818b31d63a
Reviewed-on: http://review.couchbase.org/91777
Reviewed-by: Sundar Sridharan <sundar@couchbase.com>
Tested-by: Sundar Sridharan <sundar@couchbase.com>