MB-35666: Forestdb: Renaming a compacted file can lead to the file getting deleted
When a file is compacted the older file is deleted when there are
no references to it. This causes an issue when the compacted file is
renamed to the original file and then backed up. When the backed up
file is opened, the original file gets deleted as the name of the
previous file for the compacted file matches the renamed file.

The fix is to not remove the old file if it is in a different dirrectory.

Change-Id: I7d3fc397768483976567774f79c49882847aba18
Reviewed-on: http://review.couchbase.org/113762
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>