MB-35323: Forestdb: Renaming a compacted file can lead to the file getting deleted
Backport of MB-35666

When a file is compacted the older file is deleted when there are
no references to it. This causes an issue when the compacted file is
renamed to the original file and then backed up. When the backed up
file is opened, the original file gets deleted as the name of the
previous file for the compacted file matches the renamed file.

The fix is to not remove the old file if it is in a different directory.

Change-Id: Icdbe7e9506cb932190e7adec30f9ad9881dd9d80
Reviewed-on: http://review.couchbase.org/113959
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>
Tested-by: Srinath Duvuru <srinath.duvuru@couchbase.com>