MB-36273: BP 5.5.x Forestdb reserve bitmap revision number not set correctly
Original MB-36248

The stale blocks used by block reuse are managed using two bitmaps.
One is the current bit map and the other is the reserve bitmap. The
reserve bitmap is switced with the main bitmap when mainbitmap is
exhausted. The bitmaps have revision numebrs associated with them and
must be set correctly for the block reuse to work. The revision number
of the reserve bit map will be one more than the main bit map's version.
In the case the bitamps are read from the file on disk, the version
number must be initialized correctly. The reserve bitmap's revision
number was not set correcty and the block resue failed.

The fix is to set the revision number correctly. Also, there is a
corruption issue with cached btree nodes when block reuse is used.
The cached blocks remain in cache even after the blocks are reused.
The fix is to disable the b-tree node cache.

Change-Id: I348046d7dd8ff6920d16ce78797128304762ad5c
Reviewed-on: http://review.couchbase.org/115646
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>
Reviewed-on: http://review.couchbase.org/115781
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Srinath Duvuru <srinath.duvuru@couchbase.com>