MB-36342: Intermittent Rebalance Failure
The fix for MB-34678 added cheeck to get the memory limit for a containerized
environment. The value is obtained from  /sys/fs/cgroup/memory/memory.limit_in_bytes
The value when not set is initialized to 9223372036854771712 (max number of 4k blocks
represented by a 64 bit int). But, it looks like in some case, it may be set to
18446744073709551615 (max unsined int). As the value was being read as an int it was
interpreted as -1. The ram size was set incorrectly. The fix is to read it as unsigned int.

Change-Id: Ifd85ad90eb65f9c382438bb2b41cdc9c67c3105c
Reviewed-on: http://review.couchbase.org/116189
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>