MB-45882: [Windows] - Indexes are missing
After doing offline upgrade from 6.6.2 -> 7.0.0 indexes are missing.

The issue is in the function "compactor_get_actual_filename". The
call to FindFirstFiles uses the filename and not the full path. This
may result in not getting the list of files in the directory. If no
file is found a new file is created and used. That will not have the
data that was moved to the new file causing the missing indexes problem.

The reason is when indemetadata

Change-Id: I659ba2454a4136c29b28fe61768513313db390ff
Reviewed-on: http://review.couchbase.org/c/forestdb/+/152927
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>