UBSan: Avoid accessing nullptr in filemgr_dirty_update_remove_node
filemgr_dirty_update_remove_node can perform address calculation on a
null 'node' argument with the expression "&node->avl". This is
technically undefined behaviour and hence is flagged by UBSan.

In practice, the expression "&node->avl" is equivant to "node" given
"avl" is the first member of the struct, and hence the subsequent
nullptr check in avl_remove causes no actual nullptr dereference to
occur:

    void avl_remove(struct avl_tree *tree,
		    struct avl_node *node)
    {
	// not found
	if (node == NULL) return;
	...

However, to keep UBSan happy and prevent it aborting the process if
this scenario is encoutered, add an explicit nullptr check and early
return in filemgr_dirty_update_remove_node.

Change-Id: I1c92f7d2f6bc50d460247ebd5482407201d6b45b
Reviewed-on: http://review.couchbase.org/c/forestdb/+/159313
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>