MB-47166: Fix undefined behaviour in iterator_functional_test
As identified during testing for AArch64 support,
iterator_functional_test has a number of instances of undefined
behaviour:

1. iterator_functional_test.cc:776:16: runtime error: null pointer passed as argument 1, which is declared to never be null

The test is trying to use memcmp (as used in implementaiton of
TEST_CMP macro on a nullptr). In this case the length should be zero
so just check that explicitly.

2. iterator_functional_test.cc:1206:29: runtime error: load of address 0xffffe75d46f8 with insufficient space for an object of type 'int'

The test is trying to access the 31st element in a 30-element
array. This actually causes the unit-tests to reliably fail on AArch64
- I assume because we end up with a different undefined value compared
to what x86 (normally) sees and hence it ends up failing a chec
on how many iterations have been performed.

Fix the for() loop termination so it doesn't access off the end.

With this patch AArch64 passes all unit tests.

Change-Id: I31184f0fe8c781575c61864cda2bdfc6c3457e1e
Reviewed-on: http://review.couchbase.org/c/forestdb/+/159305
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>