MB-57466: Read offset exceeds the file's current offset error
This is an error reported by forestdb. With block reuse,
a previous stale header block can get reclaimed and reused
for incoming writes. Later, when collecting stale blocks
to be reclaimed, this stale header can get accessed
when walking back to get previous header. This returns an
error and the error is not handled and ends up using
an unfilled header buffer. If the block id in the un-initialized
 header buffer is zero, then the stale block collection
ends without reading the block. If the blockid in the
header buffer is non zero, then the read continues and
tries to read an invalid offset and reports the above error.

The fix is to zero out the header buffer after allcoaction.
This will guarantee the stale header reuse is correctly handled.

Change-Id: I9cefe115d67343e87493ea696d9bf0b35b5f4a27
Reviewed-on: https://review.couchbase.org/c/forestdb/+/195240
Reviewed-by: <saptarshi.sen@couchbase.com>
Tested-by: Srinath Duvuru <srinath.duvuru@couchbase.com>
Reviewed-by: Srinath Duvuru <srinath.duvuru@couchbase.com>